refactor: Move middleware to src/Host with namespace updates
